Cleanup code

+++ FinancialAnalysis.Logic/Mail.cs
@@ -1,12 +1,8 @@
-﻿using FinancialAnalysis.Models.Mail;
-using System;
-using System.Collections.Generic;
+﻿using System;
 using System.IO;
-using System.Linq;
 using System.Net;
 using System.Net.Mail;
-using System.Text;
-using System.Threading.Tasks;
+using FinancialAnalysis.Models.Mail;
 
 namespace FinancialAnalysis.Logic
 {
@@ -18,30 +14,27 @@ namespace FinancialAnalysis.Logic
             var password = "YrktvGZZu2M4eGpP";
             var server = "w011d665.kasserver.com";
 
-            MailMessage message = new MailMessage(mailConfiguration.Address, mailData.To)
+            var message = new MailMessage(mailConfiguration.Address, mailData.To)
             {
                 Subject = mailData.Subject,
-                Body = mailData.Body,
+                Body = mailData.Body
             };
 
             if (!string.IsNullOrEmpty(mailData.AttachmentPath))
             {
-                System.Net.Mail.Attachment attachment;
-                attachment = new System.Net.Mail.Attachment(mailData.AttachmentPath);
+                Attachment attachment;
+                attachment = new Attachment(mailData.AttachmentPath);
                 message.Attachments.Add(attachment);
             }
 
             try
             {
-                SmtpClient client = new SmtpClient(server);
+                var client = new SmtpClient(server);
                 client.Credentials = new NetworkCredential(user, password);
                 client.Send(message);
                 message.Attachments.Dispose();
 
-                if (!string.IsNullOrEmpty(mailData.AttachmentPath))
-                {
-                    File.Delete(mailData.AttachmentPath);
-                }
+                if (!string.IsNullOrEmpty(mailData.AttachmentPath)) File.Delete(mailData.AttachmentPath);
             }
             catch (Exception ex)
             {
